Moved get testByThemeId to themes API
